Implement basis function evaluation.

+++ include/femto/mesh/mesh.h
@@ -8,7 +8,7 @@
 struct Fto2DMesh
 {
     double *nodes; /* C order 2D array num_nodes x 2 */
-    double *triangles; /* C order 2D array num_triangles x 3, where each element triangles[i,j] contains the index (into
+    int *triangles; /* C order 2D array num_triangles x 3, where each element triangles[i,j] contains the index (into
                         * the nodes array) of the j'th node belonging to the i'th triangle */
     int num_nodes;
     int num_triangles;
